Reorganize server routes
